fixed requested changes + linting
