+++ backend/src/errors/service.ts
@@ -23,6 +23,8 @@ const ERROR_GETTING_EVENT_DATA = "There was an error retrieving the calendly eve
 const ERROR_DELETING_EVENT = "There was an error deleting a calendly event";
 const INVALID_ROLE_WAS_FOUND = "Allowed user roles for this context is mentor and mentee only";
 
+const ERROR_SENDING_NOTIFICATION = "There was an error sending the notification.";
+
 export class ServiceError extends CustomError {
   static IMAGE_NOT_SAVED = new ServiceError(0, 404, IMAGE_NOT_SAVED);
 
@@ -50,7 +52,9 @@ export class ServiceError extends CustomError {
 
   static MENTEE_WAS_NOT_SAVED = new ServiceError(12, 404, MENTEE_WAS_NOT_SAVED);
 
-  static INVALID_ROLE_WAS_FOUND = new ServiceError(11, 404, INVALID_ROLE_WAS_FOUND);
+  static INVALID_ROLE_WAS_FOUND = new ServiceError(13, 404, INVALID_ROLE_WAS_FOUND);
 
-  static ERROR_DELETING_EVENT = new ServiceError(7, 404, ERROR_DELETING_EVENT);
+  static ERROR_DELETING_EVENT = new ServiceError(14, 404, ERROR_DELETING_EVENT);
+
+  static ERROR_SENDING_NOTIFICATION = new ServiceError(15, 404, ERROR_SENDING_NOTIFICATION);
 }
